+++ demo/components/universal/SubRoutesNav/SubRoutesNav.jsx
@@ -1,7 +1,7 @@
 // @flow
 import React from 'react'
 import { Link } from 'react-router-dom'
-import { Dropdown, Icon } from 'componentry'
+import { Dropdown } from 'componentry'
 
 type Props = {
   label: string,
@@ -10,9 +10,7 @@ type Props = {
 
 const SubRoutesNav = ({ subRoutes, label }: Props) => (
   <Dropdown as="nav">
-    <Dropdown.Trigger color="link">
-      {label} <Icon id="chevron" />
-    </Dropdown.Trigger>
+    <Dropdown.Trigger color="link">{label}</Dropdown.Trigger>
     <Dropdown.Content className="dropdown-menu-right">
       {subRoutes.map(routeTo => (
         <Dropdown.Item as={Link} to={routeTo} key={routeTo.id}>